# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/spl": {
      "get": {
        "description": "This method returns token data",
        "operationId": "SolanaTokenController_getData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolanaTokenDataEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get token data",
        "tags": [
          "spl-token"
        ]
      },
      "post": {
        "description": "Creates PumpFun token",
        "operationId": "SolanaTokenController_createToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePumpFunTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create SPL token",
        "tags": [
          "spl-token"
        ]
      }
    },
    "/spl/buy-all": {
      "post": {
        "description": "This method buys SPL tokens from all accounts, based on provided data",
        "operationId": "SolanaTokenController_buyAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyAllSPLDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SPL tokens buy started!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buy SPL tokens",
        "tags": [
          "spl-token"
        ]
      }
    },
    "/spl/transfer-all": {
      "put": {
        "description": "This method transfers all SPL tokens to defined address. EVERYONE WILL BE KNOW WHO IS THE RECIPIENT AND WHAT HAPPENED",
        "operationId": "SolanaTokenController_transferAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferAllDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SPL tokens transfer started!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transfer all SPL tokens to defined address",
        "tags": [
          "spl-token"
        ]
      }
    },
    "/spl/sell-all": {
      "delete": {
        "description": "This method sells all SPL tokens, based on provided data",
        "operationId": "SolanaTokenController_sellAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellAllSPLDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SPL tokens sell started!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsResponseEntity"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sell all SPL tokens",
        "tags": [
          "spl-token"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bundler API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SolanaTokenDataEntity": {
        "type": "object",
        "properties": {
          "mint": {
            "type": "string",
            "description": "The public key of the token",
            "example": "So11111111111111111111111111111111111111112"
          },
          "name": {
            "type": "string",
            "description": "The name of the token",
            "example": "Solana"
          },
          "symbol": {
            "type": "string",
            "description": "The symbol of the token",
            "example": "SOL"
          },
          "description": {
            "type": "string",
            "description": "The description of the token",
            "example": "Solana is a blockchain that allows you to build and run applications on its network."
          },
          "image_uri": {
            "type": "string",
            "description": "The image of the token",
            "example": "https://solana.com/wp-content/uploads/2022/01/solana-logo.png"
          },
          "metadata_uri": {
            "type": "string",
            "description": "The metadata of the token",
            "example": "https://solana.com/wp-content/uploads/2022/01/solana-logo.png"
          },
          "twitter": {
            "type": "object",
            "description": "The Twitter handle of the token",
            "example": "@solana",
            "nullable": true
          },
          "telegram": {
            "type": "object",
            "description": "The Telegram group of the token",
            "example": "https://t.me/solana",
            "nullable": true
          },
          "bonding_curve": {
            "type": "string",
            "description": "The bonding curve address of the token",
            "example": "6QzTnHVtirQoFESMfruoErj63VdQ1Mwug8ku7cozE7zh"
          },
          "associated_bonding_curve": {
            "type": "string",
            "description": "The associated bonding curve address of the token",
            "example": "9W5RLtqDUPUFRHBBqAtULufGLEmjUZeS1wLKUFXhNSfg"
          },
          "creator": {
            "type": "string",
            "description": "The creator's address",
            "example": "CJCPJNLB9R3nWcetzZ3XqUNNpgG8FT1ud37zMYNcKZn4"
          },
          "created_timestamp": {
            "type": "number",
            "description": "The creation timestamp",
            "example": 1718270774645
          },
          "raydium_pool": {
            "type": "object",
            "description": "The Raydium pool address",
            "example": "RaydiumPoolAddress",
            "nullable": true
          },
          "complete": {
            "type": "boolean",
            "description": "Indicates if the token data is complete",
            "example": false
          },
          "virtual_sol_reserves": {
            "type": "number",
            "description": "The virtual Solana reserves",
            "example": 30100362525
          },
          "virtual_token_reserves": {
            "type": "number",
            "description": "The virtual token reserves",
            "example": 1069422363776860
          },
          "total_supply": {
            "type": "number",
            "description": "The total supply of the token",
            "example": 1000000000000000
          },
          "website": {
            "type": "object",
            "description": "The website of the token",
            "example": "https://solana.com",
            "nullable": true
          },
          "show_name": {
            "type": "boolean",
            "description": "Indicates if the token name should be shown",
            "example": true
          },
          "king_of_the_hill_timestamp": {
            "type": "object",
            "description": "The King of the Hill timestamp",
            "example": 1718270774645,
            "nullable": true
          },
          "market_cap": {
            "type": "number",
            "description": "The market cap of the token",
            "example": 28.146374664
          },
          "reply_count": {
            "type": "number",
            "description": "The reply count of the token",
            "example": 2
          },
          "last_reply": {
            "type": "number",
            "description": "The last reply timestamp",
            "example": 1719166751503
          },
          "nsfw": {
            "type": "boolean",
            "description": "Indicates if the token is NSFW",
            "example": false
          },
          "market_id": {
            "type": "object",
            "description": "The market ID",
            "example": "MarketID",
            "nullable": true
          },
          "inverted": {
            "type": "object",
            "description": "Indicates if the token is inverted",
            "example": false,
            "nullable": true
          },
          "username": {
            "type": "string",
            "description": "The username of the creator",
            "example": "ASATIANI"
          },
          "profile_image": {
            "type": "string",
            "description": "The profile image URI",
            "example": "https://cf-ipfs.com/ipfs/QmZjkUUwxnMJqJ5q2Zre7iNUgS9Puew3Z6tazsLqcErKBL"
          },
          "usd_market_cap": {
            "type": "number",
            "description": "The USD market cap of the token",
            "example": 3563.61249620904
          },
          "solana_price_per_token": {
            "type": "number",
            "description": "The solana price per token",
            "example": 0.001
          },
          "usd_price_per_token": {
            "type": "number",
            "description": "The USD price per token",
            "example": 0.001
          }
        },
        "required": [
          "mint",
          "name",
          "symbol",
          "description",
          "image_uri",
          "metadata_uri",
          "twitter",
          "telegram",
          "bonding_curve",
          "associated_bonding_curve",
          "creator",
          "created_timestamp",
          "raydium_pool",
          "complete",
          "virtual_sol_reserves",
          "virtual_token_reserves",
          "total_supply",
          "website",
          "show_name",
          "king_of_the_hill_timestamp",
          "market_cap",
          "reply_count",
          "last_reply",
          "nsfw",
          "market_id",
          "inverted",
          "username",
          "profile_image",
          "usd_market_cap",
          "solana_price_per_token",
          "usd_price_per_token"
        ]
      },
      "TokenBuyerDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Address if of the token buyer"
          },
          "solAmount": {
            "type": "number",
            "description": "Amount of solana to spend"
          }
        },
        "required": [
          "address",
          "solAmount"
        ]
      },
      "CreatePumpFunTokenDto": {
        "type": "object",
        "properties": {
          "owner": {
            "type": "string",
            "description": "Public key of of account that will be used for creating token"
          },
          "buyers": {
            "description": "Token buyer",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenBuyerDto"
            }
          },
          "name": {
            "type": "string",
            "description": "Name of the token"
          },
          "symbol": {
            "type": "string",
            "description": "Symbol of the token"
          },
          "description": {
            "type": "string",
            "description": "Description of the token"
          },
          "twitter": {
            "type": "string",
            "description": "Twitter"
          },
          "telegram": {
            "type": "string",
            "description": "Telegram"
          },
          "website": {
            "type": "string",
            "description": "Website"
          },
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "owner",
          "buyers",
          "name",
          "symbol",
          "description",
          "file"
        ]
      },
      "TransactionsResponseEntity": {
        "type": "object",
        "properties": {
          "txHashes": {
            "description": "The transaction hashes of the transfers",
            "example": [
              "txHash1",
              "txHash2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "txHashes"
        ]
      },
      "BuyAllSPLDto": {
        "type": "object",
        "properties": {
          "keepSolanaAmount": {
            "type": "number",
            "description": "The amount of solana that will be left on the accounts",
            "example": 0.01
          },
          "percent": {
            "type": "number",
            "description": "The percentage of the solana balance that will be bought",
            "example": 100
          },
          "slippagePercent": {
            "type": "number",
            "description": "The slippage percentage",
            "example": 1
          },
          "priorityMicroLamptorsFee": {
            "type": "number",
            "description": "The priority lamports fee",
            "example": 0
          }
        },
        "required": [
          "percent"
        ]
      },
      "TransferAllDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "The address to transfer SPL tokens to"
          },
          "priorityMicroLamptorsFee": {
            "type": "number",
            "description": "The priority lamports fee",
            "example": 40000
          }
        },
        "required": [
          "address"
        ]
      },
      "SellAllSPLDto": {
        "type": "object",
        "properties": {
          "keepAmount": {
            "type": "number",
            "description": "The amount of SPL tokens that will be left on the accounts",
            "example": 0
          },
          "percent": {
            "type": "number",
            "description": "The percentage of how much for each account will be sold",
            "example": 100
          },
          "slippagePercent": {
            "type": "number",
            "description": "The slippage percentage",
            "example": 1
          },
          "priorityMicroLamptorsFee": {
            "type": "number",
            "description": "The priority lamports fee",
            "example": 0
          }
        },
        "required": [
          "percent"
        ]
      },
...
    }
  }
}